remove old server
