--- external/a2a/src/server2/plugin.ts
@@ -1,111 +1,0 @@
-import { randomUUID } from 'node:crypto';
-
-import { AgentCard, Message, Task } from '@a2a-js/sdk';
-import { A2ARequestHandler, AgentExecutor, DefaultRequestHandler, ExecutionEventBus, InMemoryTaskStore, RequestContext, TaskStore } from '@a2a-js/sdk/server';
-import { A2AExpressApp } from '@a2a-js/sdk/server/express';
-import express from 'express';
-
-import { Dependency, EmitPluginEvent, Event, HttpPlugin, IPlugin, Plugin } from '@microsoft/teams.apps';
-
-
-interface IA2APluginOptions {
-  /**
-   * The agent card to be used for the A2A plugin.
-   */
-  agentCard: AgentCard;
-
-  /**
-   * Path to the A2A server
-   * @default '/a2a'
-   */
-  path?: `/${string}`;
-
-  /**
-   * taskStore which stores the tasks that are sent to the agent
-   * or that the agent sends. If not provided, the App's storage will be used.
-   */
-  taskStore?: TaskStore;
-
-  /**
-  * For a completely custom executor, you may provide your own executor that will
-  * get executed whenever the a2a agent is InMemoryTaskStore
-  */
-  agentExecutor?: AgentExecutor;
-}
-
-export type Respond = (message: string | Message | Task) => Promise<void>;
-export type A2AEvents = {
-  'a2a:message': {
-    requestContext: RequestContext;
-    respond: Respond;
-    publishUpdate: ExecutionEventBus['publish']
-  }
-};
-
-@Plugin({
-  name: 'a2a',
-  description: 'A2A Server Plugin',
-  version: '0.3.0'
-})
-export class A2APlugin implements IPlugin {
-  @Event('custom')
-  protected readonly emit!: EmitPluginEvent<A2AEvents>;
-
-  @Dependency()
-  protected readonly _httpPlugin!: HttpPlugin;
-  __eventType!: A2AEvents;
-
-  protected _card: AgentCard;
-  protected _path: string;
-  protected _taskStore: TaskStore;
-  protected _customExecutor?: AgentExecutor;
-
-  constructor(options: IA2APluginOptions) {
-    this._card = options.agentCard;
-    this._path = options.path ?? '/a2a';
-    this._taskStore = options.taskStore ?? new InMemoryTaskStore();
-    this._customExecutor = options.agentExecutor;
-  }
-
-  onInit() {
-    const a2aExpressApp = new A2AExpressApp(this._setupRequestHandler());
-    const expressApp = express();
-    a2aExpressApp.setupRoutes(expressApp);
-    this._httpPlugin.use(expressApp);
-  }
-
-  _setupExecutor() {
-    const executor: AgentExecutor = {
-      execute: async (requestContext, eventBus) => {
-        const ctx: A2AEvents['a2a:message'] = {
-          requestContext,
-          respond: async (message) => {
-            let responseMessage: Message | Task;
-            if (typeof message === 'string') {
-              responseMessage = {
-                kind: 'message',
-                messageId: randomUUID(),
-                role: 'agent',
-                parts: [{ kind: 'text', text: message }],
-                // Associate the response with the incoming request's context.
-                contextId: requestContext.contextId,
-              };
-            } else {
-              responseMessage = message;
-            }
-            eventBus.publish(responseMessage);
-            eventBus.finished();
-          },
-          publishUpdate: eventBus.publish.bind(eventBus)
-        };
-        this.emit('a2a:message', ctx);
-      },
-      cancelTask: async () => { }
-    };
-    return executor;
-  }
-
-  _setupRequestHandler(): A2ARequestHandler {
-    return new DefaultRequestHandler(this._card, this._taskStore, this._customExecutor ?? this._setupExecutor());
-  }
-}
